+++ client/assets/js/category.js
@@ -1,0 +1,3 @@
+const { renderJournalsToPage } = require('./lib/handlers');
+
+document.addEventListener('DOMContentLoaded', renderJournalsToPage);