+++ client/assets/js/lib/handlers.js
@@ -1,0 +1,11 @@
+const { createAllJournals } = require('./helpers');
+
+async function renderJournalsToPage() {
+	const journals = await createAllJournals();
+	const main = document.querySelector('main');
+	journals.forEach((journal) => {
+		main.append(journal);
+	});
+}
+
+module.exports = { renderJournalsToPage };