+++ changelog.md
@@ -6,8 +6,8 @@
 [X] Features
 [X] Pages list
 [X] Api Endpoints list
-[] Make TODO on notion - write initial todo
-[] Make Organisation - Go with path of least resistance (morris/ram)
+[X] Make TODO on notion - write initial todo
+[X] Make Organisation - Go with path of least resistance (morris/ram)
 
 ## The plan
 
@@ -79,20 +79,47 @@ Pages
 
 - `/`
   - GET `Hello World`
-- `/journels`
+- `/journals`
   - POST
   - GET
-- /journels/:journeyId
+- `/journals/categories/`
+  - GET
+- `/journals/categories/:category`
+  - GET
+- `/journals/:journalId`
   - PUT
   - DELETE
   - GET (maybe)
-- /journels/:journeyId/:emoji
+- `/journals/:journalId/:emoji`
   - PUT
-- /journels/:journeyId/comments
+- `/journals/:journalId/comments`
   - GET
   - POST
-- /journels/:journeyId/comments/:commentId
+- `/journals/:journalId/comments/:commentId`
   - PUT
   - DELETE
-- /categories/
-  - GET
+
+## Changes
+
+## Akash
+
+- Create frontend wireframe with Figma paired with Rafika
+- In journal.js added a function to facilitate the adding of new journal entries, in the models folder.
+- 06/07/21 I have added all the neccesary methods to the journals model and finished it off, now it is completely finished.
+- Post data validated
+
+### Frontend
+
+### Backend
+
+- stub basic architecture
+- create basic api, GET `/`
+- create journels router and mount on `/journals/`
+- stub all api routes on journals router
+- abstract `/journals/categories` to router
+- abstract `/journals/:journalId/comments` to router
+- create `Comment` model
+- abstract JSON file writing logic
+- POST validations for client and server
+
+### Misc